Added capability to shoot rings.
Once right trigger is pressed on driver 2's controller (in the shooting phase), the servo which feeds rings into the shooter pushes a ring into the flywheel. Should have capability to shoot 3 rings in 0.9 seconds. Uses assumption that the feeder is a fixed (non CR) servo with 2 or 4 bar linkage.