Implemented automatic aiming
Uses AlignLeft/AlignRight to drive robot into ideal position for shooting. Later, color sensors might be included as to drive to optimal shooting position relative to the shooting line. On gamepad one, left on the D-pad will align the robot on the left side of the field, and right on the D-pad will do the opposite.